Added Event for MoneyPayment
I added a Payment Event, called, when a player pay another player with "/money pay <Name> <amount> <Message>".

The event is called MoneyTransactionEvent. You can set it to cancelled, if the payment should stopped.
This Event is also fired, if the transaction is stopped, caused by the player dont have the money, there is no acc or the player want to pay himself.

At least I removed the Permissions dependency, cause this plugin is (like this) dead.